Added the logdata modifier

+++ engine/config-parser.c
@@ -1,5 +1,5 @@
 
-#line 1 "config-parser.rl"
+#line 1 "../../ib/engine/config-parser.rl"
 /*****************************************************************************
  * Licensed to Qualys, Inc. (QUALYS) under one or more
  * contributor license agreements.  See the NOTICE file distributed with
@@ -44,7 +44,7 @@
  * Finite state machine type.
  * @internal
  * Contains state information for Ragel's parser.
- * Many of these values and names come from the Ragel documentation, section
+ * Many of these values and names come from the Ragel documentation, section 
  * 5.1 Variable Used by Ragel. p53 of The Ragel Guide 6.7 found at
  * http://www.complang.org/ragel/ragel-guide-6.7.pdf
  */
@@ -95,163 +95,163 @@ static char* alloc_cpy_marked_string(const char *fpc_mark,
 }
 
 
-#line 199 "config-parser.rl"
+#line 199 "../../ib/engine/config-parser.rl"
 
 
 
-#line 103 "config-parser.c"
+#line 103 "../../ib/engine/config-parser.c"
 static const char _ironbee_config_actions[] = {
-	0, 1, 0, 1, 1, 1, 11, 1,
-	20, 1, 22, 1, 26, 1, 31, 1,
-	32, 1, 33, 1, 35, 1, 36, 1,
-	37, 2, 1, 24, 2, 1, 28, 2,
-	2, 5, 2, 2, 15, 2, 3, 19,
-	2, 4, 32, 2, 4, 34, 2, 5,
-	14, 2, 6, 22, 2, 6, 23, 2,
-	7, 18, 2, 8, 27, 2, 9, 10,
-	2, 12, 13, 2, 12, 17, 2, 12,
-	21, 2, 12, 25, 2, 12, 29, 2,
-	12, 30, 3, 2, 5, 14, 3, 2,
-	5, 15, 3, 2, 5, 16, 3, 12,
-	0, 13, 3, 12, 0, 17, 3, 12,
-	0, 21, 3, 12, 0, 25, 3, 12,
+	0, 1, 0, 1, 1, 1, 11, 1, 
+	20, 1, 22, 1, 26, 1, 31, 1, 
+	32, 1, 33, 1, 35, 1, 36, 1, 
+	37, 2, 1, 24, 2, 1, 28, 2, 
+	2, 5, 2, 2, 15, 2, 3, 19, 
+	2, 4, 32, 2, 4, 34, 2, 5, 
+	14, 2, 6, 22, 2, 6, 23, 2, 
+	7, 18, 2, 8, 27, 2, 9, 10, 
+	2, 12, 13, 2, 12, 17, 2, 12, 
+	21, 2, 12, 25, 2, 12, 29, 2, 
+	12, 30, 3, 2, 5, 14, 3, 2, 
+	5, 15, 3, 2, 5, 16, 3, 12, 
+	0, 13, 3, 12, 0, 17, 3, 12, 
+	0, 21, 3, 12, 0, 25, 3, 12, 
 	0, 29
 };
 
 static const unsigned char _ironbee_config_key_offsets[] = {
-	0, 0, 0, 0, 9, 11, 11, 11,
-	19, 21, 21, 21, 30, 30, 38, 42,
-	43, 52, 59, 68, 75, 76, 77, 86,
-	93, 100, 100, 108, 115, 115, 124, 131,
+	0, 0, 0, 0, 9, 11, 11, 11, 
+	19, 21, 21, 21, 30, 30, 38, 42, 
+	43, 52, 59, 68, 75, 76, 77, 86, 
+	93, 100, 100, 108, 115, 115, 124, 131, 
 	138, 146
 };
 
 static const char _ironbee_config_trans_keys[] = {
-	9, 10, 13, 32, 34, 35, 60, 62,
-	92, 34, 92, 9, 10, 32, 34, 35,
-	60, 62, 92, 34, 92, 9, 10, 13,
-	32, 60, 62, 92, 34, 35, 9, 10,
-	32, 60, 62, 92, 34, 35, 9, 10,
-	13, 32, 10, 9, 10, 13, 32, 34,
-	35, 60, 62, 92, 32, 34, 60, 62,
-	92, 9, 10, 9, 10, 13, 32, 34,
-	35, 60, 62, 92, 9, 10, 32, 34,
-	60, 62, 92, 10, 47, 9, 10, 13,
-	32, 34, 35, 60, 62, 92, 32, 34,
-	60, 62, 92, 9, 10, 9, 10, 32,
-	34, 60, 62, 92, 9, 10, 32, 34,
-	35, 60, 62, 92, 32, 34, 60, 62,
-	92, 9, 10, 9, 10, 13, 32, 60,
-	62, 92, 34, 35, 32, 34, 60, 62,
-	92, 9, 10, 9, 10, 32, 34, 60,
-	62, 92, 9, 10, 32, 60, 62, 92,
-	34, 35, 32, 34, 60, 62, 92, 9,
+	9, 10, 13, 32, 34, 35, 60, 62, 
+	92, 34, 92, 9, 10, 32, 34, 35, 
+	60, 62, 92, 34, 92, 9, 10, 13, 
+	32, 60, 62, 92, 34, 35, 9, 10, 
+	32, 60, 62, 92, 34, 35, 9, 10, 
+	13, 32, 10, 9, 10, 13, 32, 34, 
+	35, 60, 62, 92, 32, 34, 60, 62, 
+	92, 9, 10, 9, 10, 13, 32, 34, 
+	35, 60, 62, 92, 9, 10, 32, 34, 
+	60, 62, 92, 10, 47, 9, 10, 13, 
+	32, 34, 35, 60, 62, 92, 32, 34, 
+	60, 62, 92, 9, 10, 9, 10, 32, 
+	34, 60, 62, 92, 9, 10, 32, 34, 
+	35, 60, 62, 92, 32, 34, 60, 62, 
+	92, 9, 10, 9, 10, 13, 32, 60, 
+	62, 92, 34, 35, 32, 34, 60, 62, 
+	92, 9, 10, 9, 10, 32, 34, 60, 
+	62, 92, 9, 10, 32, 60, 62, 92, 
+	34, 35, 32, 34, 60, 62, 92, 9, 
 	10, 0
 };
 
 static const char _ironbee_config_single_lengths[] = {
-	0, 0, 0, 9, 2, 0, 0, 8,
-	2, 0, 0, 7, 0, 6, 4, 1,
-	9, 5, 9, 7, 1, 1, 9, 5,
-	7, 0, 8, 5, 0, 7, 5, 7,
+	0, 0, 0, 9, 2, 0, 0, 8, 
+	2, 0, 0, 7, 0, 6, 4, 1, 
+	9, 5, 9, 7, 1, 1, 9, 5, 
+	7, 0, 8, 5, 0, 7, 5, 7, 
 	6, 5
 };
 
 static const char _ironbee_config_range_lengths[] = {
-	0, 0, 0, 0, 0, 0, 0, 0,
-	0, 0, 0, 1, 0, 1, 0, 0,
-	0, 1, 0, 0, 0, 0, 0, 1,
-	0, 0, 0, 1, 0, 1, 1, 0,
+	0, 0, 0, 0, 0, 0, 0, 0, 
+	0, 0, 0, 1, 0, 1, 0, 0, 
+	0, 1, 0, 0, 0, 0, 0, 1, 
+	0, 0, 0, 1, 0, 1, 1, 0, 
 	1, 1
 };
 
 static const unsigned char _ironbee_config_index_offsets[] = {
-	0, 0, 1, 2, 12, 15, 16, 17,
-	26, 29, 30, 31, 40, 41, 49, 54,
-	56, 66, 73, 83, 91, 93, 95, 105,
-	112, 120, 121, 130, 137, 138, 147, 154,
+	0, 0, 1, 2, 12, 15, 16, 17, 
+	26, 29, 30, 31, 40, 41, 49, 54, 
+	56, 66, 73, 83, 91, 93, 95, 105, 
+	112, 120, 121, 130, 137, 138, 147, 154, 
 	162, 170
 };
 
 static const char _ironbee_config_indicies[] = {
-	1, 3, 5, 6, 7, 5, 8, 9,
-	9, 9, 10, 4, 12, 13, 11, 11,
-	15, 17, 9, 17, 18, 9, 9, 19,
-	20, 16, 22, 23, 21, 21, 25, 27,
-	28, 29, 27, 30, 30, 31, 30, 26,
-	33, 35, 30, 35, 30, 36, 37, 30,
-	34, 36, 38, 39, 36, 30, 38, 30,
-	41, 42, 43, 41, 9, 44, 45, 9,
-	46, 40, 47, 47, 47, 47, 48, 47,
-	1, 41, 42, 43, 41, 49, 44, 49,
-	49, 46, 40, 47, 50, 47, 47, 47,
-	47, 48, 1, 51, 44, 53, 52, 5,
-	6, 7, 5, 8, 9, 9, 9, 10,
-	4, 55, 55, 55, 55, 56, 55, 3,
-	55, 57, 55, 55, 55, 55, 56, 3,
-	55, 17, 9, 17, 18, 9, 9, 19,
-	20, 16, 58, 58, 58, 58, 59, 58,
-	15, 58, 27, 28, 29, 27, 30, 30,
-	31, 30, 26, 60, 60, 60, 60, 61,
-	60, 25, 60, 62, 60, 60, 60, 60,
-	61, 25, 35, 30, 35, 30, 36, 37,
-	30, 34, 63, 63, 63, 63, 64, 63,
+	1, 3, 5, 6, 7, 5, 8, 9, 
+	9, 9, 10, 4, 12, 13, 11, 11, 
+	15, 17, 9, 17, 18, 9, 9, 19, 
+	20, 16, 22, 23, 21, 21, 25, 27, 
+	28, 29, 27, 30, 30, 31, 30, 26, 
+	33, 35, 30, 35, 30, 36, 37, 30, 
+	34, 36, 38, 39, 36, 30, 38, 30, 
+	41, 42, 43, 41, 9, 44, 45, 9, 
+	46, 40, 47, 47, 47, 47, 48, 47, 
+	1, 41, 42, 43, 41, 49, 44, 49, 
+	49, 46, 40, 47, 50, 47, 47, 47, 
+	47, 48, 1, 51, 44, 53, 52, 5, 
+	6, 7, 5, 8, 9, 9, 9, 10, 
+	4, 55, 55, 55, 55, 56, 55, 3, 
+	55, 57, 55, 55, 55, 55, 56, 3, 
+	55, 17, 9, 17, 18, 9, 9, 19, 
+	20, 16, 58, 58, 58, 58, 59, 58, 
+	15, 58, 27, 28, 29, 27, 30, 30, 
+	31, 30, 26, 60, 60, 60, 60, 61, 
+	60, 25, 60, 62, 60, 60, 60, 60, 
+	61, 25, 35, 30, 35, 30, 36, 37, 
+	30, 34, 63, 63, 63, 63, 64, 63, 
 	33, 0
 };
 
 static const char _ironbee_config_trans_targs[] = {
-	16, 17, 22, 23, 23, 3, 22, 24,
-	4, 0, 2, 4, 25, 5, 26, 27,
-	27, 7, 8, 26, 6, 8, 28, 9,
-	29, 30, 30, 11, 29, 31, 0, 10,
-	32, 33, 33, 13, 14, 12, 32, 15,
-	17, 18, 16, 19, 20, 21, 1, 16,
-	1, 16, 16, 16, 16, 16, 22, 22,
-	2, 22, 26, 6, 29, 10, 29, 32,
+	16, 17, 22, 23, 23, 3, 22, 24, 
+	4, 0, 2, 4, 25, 5, 26, 27, 
+	27, 7, 8, 26, 6, 8, 28, 9, 
+	29, 30, 30, 11, 29, 31, 0, 10, 
+	32, 33, 33, 13, 14, 12, 32, 15, 
+	17, 18, 16, 19, 20, 21, 1, 16, 
+	1, 16, 16, 16, 16, 16, 22, 22, 
+	2, 22, 26, 6, 29, 10, 29, 32, 
 	12
 };
 
 static const char _ironbee_config_trans_actions[] = {
-	23, 76, 90, 64, 94, 0, 46, 94,
-	1, 0, 1, 0, 0, 0, 7, 67,
-	98, 0, 1, 55, 1, 0, 0, 0,
-	25, 70, 102, 0, 9, 102, 3, 1,
-	28, 73, 106, 0, 0, 1, 11, 0,
-	110, 79, 15, 110, 0, 0, 1, 43,
-	0, 21, 40, 17, 19, 13, 86, 34,
-	0, 82, 37, 0, 52, 0, 49, 58,
+	23, 76, 90, 64, 94, 0, 46, 94, 
+	1, 0, 1, 0, 0, 0, 7, 67, 
+	98, 0, 1, 55, 1, 0, 0, 0, 
+	25, 70, 102, 0, 9, 102, 3, 1, 
+	28, 73, 106, 0, 0, 1, 11, 0, 
+	110, 79, 15, 110, 0, 0, 1, 43, 
+	0, 21, 40, 17, 19, 13, 86, 34, 
+	0, 82, 37, 0, 52, 0, 49, 58, 
 	0
 };
 
 static const char _ironbee_config_to_state_actions[] = {
-	0, 0, 0, 0, 0, 0, 0, 0,
-	0, 0, 0, 0, 0, 0, 0, 0,
-	61, 0, 0, 0, 0, 0, 61, 0,
-	0, 0, 61, 0, 0, 61, 0, 0,
+	0, 0, 0, 0, 0, 0, 0, 0, 
+	0, 0, 0, 0, 0, 0, 0, 0, 
+	61, 0, 0, 0, 0, 0, 61, 0, 
+	0, 0, 61, 0, 0, 61, 0, 0, 
 	61, 0
 };
 
 static const char _ironbee_config_from_state_actions[] = {
-	0, 0, 0, 0, 0, 0, 0, 0,
-	0, 0, 0, 0, 0, 0, 0, 0,
-	5, 0, 0, 0, 0, 0, 5, 0,
-	0, 0, 5, 0, 0, 5, 0, 0,
+	0, 0, 0, 0, 0, 0, 0, 0, 
+	0, 0, 0, 0, 0, 0, 0, 0, 
+	5, 0, 0, 0, 0, 0, 5, 0, 
+	0, 0, 5, 0, 0, 5, 0, 0, 
 	5, 0
 };
 
 static const char _ironbee_config_eof_actions[] = {
-	0, 0, 0, 31, 31, 31, 0, 0,
-	0, 0, 0, 3, 0, 3, 3, 3,
-	0, 0, 0, 0, 0, 0, 31, 0,
-	0, 0, 0, 0, 0, 0, 0, 0,
+	0, 0, 0, 31, 31, 31, 0, 0, 
+	0, 0, 0, 3, 0, 3, 3, 3, 
+	0, 0, 0, 0, 0, 0, 31, 0, 
+	0, 0, 0, 0, 0, 0, 0, 0, 
 	0, 0
 };
 
 static const unsigned char _ironbee_config_eof_trans[] = {
-	0, 1, 3, 0, 0, 0, 15, 0,
-	0, 0, 25, 0, 33, 0, 0, 0,
-	0, 48, 50, 48, 52, 53, 0, 55,
-	55, 55, 0, 59, 59, 0, 61, 61,
+	0, 1, 3, 0, 0, 0, 15, 0, 
+	0, 0, 25, 0, 33, 0, 0, 0, 
+	0, 48, 50, 48, 52, 53, 0, 55, 
+	55, 55, 0, 59, 59, 0, 61, 61, 
 	0, 64
 };
 
@@ -266,7 +266,7 @@ static const int ironbee_config_en_endblock = 32;
 static const int ironbee_config_en_main = 16;
 
 
-#line 202 "config-parser.rl"
+#line 202 "../../ib/engine/config-parser.rl"
 
 ib_status_t ib_cfgparser_ragel_parse_chunk(ib_cfgparser_t *cp,
                                            const char *buf,
@@ -317,17 +317,17 @@ ib_status_t ib_cfgparser_ragel_parse_chunk(ib_cfgparser_t *cp,
     }
 
     /* Access all ragel state variables via structure. */
+    
+#line 253 "../../ib/engine/config-parser.rl"
+    
+#line 254 "../../ib/engine/config-parser.rl"
+    
+#line 255 "../../ib/engine/config-parser.rl"
+    
+#line 256 "../../ib/engine/config-parser.rl"
 
-#line 253 "config-parser.rl"
-
-#line 254 "config-parser.rl"
-
-#line 255 "config-parser.rl"
-
-#line 256 "config-parser.rl"
-
-
-#line 331 "config-parser.c"
+    
+#line 331 "../../ib/engine/config-parser.c"
 	{
 	 fsm.cs = ironbee_config_start;
 	 fsm.top = 0;
@@ -336,9 +336,9 @@ ib_status_t ib_cfgparser_ragel_parse_chunk(ib_cfgparser_t *cp,
 	 fsm.act = 0;
 	}
 
-#line 258 "config-parser.rl"
-
-#line 342 "config-parser.c"
+#line 258 "../../ib/engine/config-parser.rl"
+    
+#line 342 "../../ib/engine/config-parser.c"
 	{
 	int _klen;
 	unsigned int _trans;
@@ -359,7 +359,7 @@ _resume:
 #line 1 "NONE"
 	{ fsm.ts = ( fsm.p);}
 	break;
-#line 363 "config-parser.c"
+#line 363 "../../ib/engine/config-parser.c"
 		}
 	}
 
@@ -381,7 +381,7 @@ _resume:
 			else if ( (*( fsm.p)) > *_mid )
 				_lower = _mid + 1;
 			else {
-				_trans += (_mid - _keys);
+				_trans += (unsigned int)(_mid - _keys);
 				goto _match;
 			}
 		}
@@ -404,7 +404,7 @@ _resume:
 			else if ( (*( fsm.p)) > _mid[1] )
 				_lower = _mid + 2;
 			else {
-				_trans += ((_mid - _keys)>>1);
+				_trans += (unsigned int)((_mid - _keys)>>1);
 				goto _match;
 			}
 		}
@@ -426,32 +426,32 @@ _eof_trans:
 		switch ( *_acts++ )
 		{
 	case 0:
-#line 99 "config-parser.rl"
+#line 99 "../../ib/engine/config-parser.rl"
 	{ mark = ( fsm.p); }
 	break;
 	case 1:
-#line 100 "config-parser.rl"
+#line 100 "../../ib/engine/config-parser.rl"
 	{
         rc = IB_EOTHER;
         ib_log_debug(ib_engine, 4, "ERROR: parser error before \"%.*s\"", (int)(( fsm.p) - mark), mark);
     }
 	break;
 	case 2:
-#line 106 "config-parser.rl"
+#line 106 "../../ib/engine/config-parser.rl"
 	{
         pval = alloc_cpy_marked_string(mark, ( fsm.p), mpcfg);
         ib_list_push(plist, pval);
     }
 	break;
 	case 3:
-#line 110 "config-parser.rl"
+#line 110 "../../ib/engine/config-parser.rl"
 	{
         pval = alloc_cpy_marked_string(mark, ( fsm.p), mpcfg);
         ib_list_push(plist, pval);
     }
 	break;
 	case 4:
-#line 116 "config-parser.rl"
+#line 116 "../../ib/engine/config-parser.rl"
 	{
         size_t namelen = (size_t)(( fsm.p) - mark);
         dirname = (char *)calloc(namelen + 1, sizeof(*dirname));
@@ -460,7 +460,7 @@ _eof_trans:
     }
 	break;
 	case 5:
-#line 122 "config-parser.rl"
+#line 122 "../../ib/engine/config-parser.rl"
 	{
         rc = ib_config_directive_process(cp, dirname, plist);
         if (rc != IB_OK) {
@@ -472,7 +472,7 @@ _eof_trans:
     }
 	break;
 	case 6:
-#line 133 "config-parser.rl"
+#line 133 "../../ib/engine/config-parser.rl"
 	{
         size_t namelen = (size_t)(( fsm.p) - mark);
         blkname = (char *)calloc(namelen + 1, sizeof(*blkname));
@@ -481,7 +481,7 @@ _eof_trans:
     }
 	break;
 	case 7:
-#line 139 "config-parser.rl"
+#line 139 "../../ib/engine/config-parser.rl"
 	{
         rc = ib_config_block_start(cp, blkname, plist);
         if (rc != IB_OK) {
@@ -490,7 +490,7 @@ _eof_trans:
     }
 	break;
 	case 8:
-#line 145 "config-parser.rl"
+#line 145 "../../ib/engine/config-parser.rl"
 	{
         blkname = (char *)cp->cur_blkname;
         rc = ib_config_block_process(cp, blkname);
@@ -508,15 +508,15 @@ _eof_trans:
 	{ fsm.te = ( fsm.p)+1;}
 	break;
 	case 13:
-#line 172 "config-parser.rl"
+#line 172 "../../ib/engine/config-parser.rl"
 	{ fsm.act = 1;}
 	break;
 	case 14:
-#line 173 "config-parser.rl"
+#line 173 "../../ib/engine/config-parser.rl"
 	{ fsm.te = ( fsm.p)+1;{ { fsm.cs =  fsm.stack[-- fsm.top]; goto _again;} }}
 	break;
 	case 15:
-#line 172 "config-parser.rl"
+#line 172 "../../ib/engine/config-parser.rl"
 	{ fsm.te = ( fsm.p);( fsm.p)--;}
 	break;
 	case 16:
@@ -532,15 +532,15 @@ _eof_trans:
 	}
 	break;
 	case 17:
-#line 177 "config-parser.rl"
+#line 177 "../../ib/engine/config-parser.rl"
 	{ fsm.act = 3;}
 	break;
 	case 18:
-#line 178 "config-parser.rl"
+#line 178 "../../ib/engine/config-parser.rl"
 	{ fsm.te = ( fsm.p)+1;{ { fsm.cs =  fsm.stack[-- fsm.top]; goto _again;} }}
 	break;
 	case 19:
-#line 177 "config-parser.rl"
+#line 177 "../../ib/engine/config-parser.rl"
 	{ fsm.te = ( fsm.p);( fsm.p)--;}
 	break;
 	case 20:
@@ -556,15 +556,15 @@ _eof_trans:
 	}
 	break;
 	case 21:
-#line 182 "config-parser.rl"
+#line 182 "../../ib/engine/config-parser.rl"
 	{ fsm.act = 5;}
 	break;
 	case 22:
-#line 183 "config-parser.rl"
+#line 183 "../../ib/engine/config-parser.rl"
 	{ fsm.te = ( fsm.p)+1;{ { fsm.cs =  fsm.stack[-- fsm.top]; goto _again;} }}
 	break;
 	case 23:
-#line 182 "config-parser.rl"
+#line 182 "../../ib/engine/config-parser.rl"
 	{ fsm.te = ( fsm.p);( fsm.p)--;{ { fsm.stack[ fsm.top++] =  fsm.cs;  fsm.cs = 26; goto _again;} }}
 	break;
 	case 24:
@@ -580,15 +580,15 @@ _eof_trans:
 	}
 	break;
 	case 25:
-#line 187 "config-parser.rl"
+#line 187 "../../ib/engine/config-parser.rl"
 	{ fsm.act = 7;}
 	break;
 	case 26:
-#line 188 "config-parser.rl"
+#line 188 "../../ib/engine/config-parser.rl"
 	{ fsm.te = ( fsm.p)+1;{ { fsm.cs =  fsm.stack[-- fsm.top]; goto _again;} }}
 	break;
 	case 27:
-#line 187 "config-parser.rl"
+#line 187 "../../ib/engine/config-parser.rl"
 	{ fsm.te = ( fsm.p);( fsm.p)--;}
 	break;
 	case 28:
@@ -604,35 +604,35 @@ _eof_trans:
 	}
 	break;
 	case 29:
-#line 193 "config-parser.rl"
+#line 193 "../../ib/engine/config-parser.rl"
 	{ fsm.act = 10;}
 	break;
 	case 30:
-#line 196 "config-parser.rl"
+#line 196 "../../ib/engine/config-parser.rl"
 	{ fsm.act = 13;}
 	break;
 	case 31:
-#line 195 "config-parser.rl"
+#line 195 "../../ib/engine/config-parser.rl"
 	{ fsm.te = ( fsm.p)+1;{ { fsm.stack[ fsm.top++] =  fsm.cs;  fsm.cs = 32; goto _again;} }}
 	break;
 	case 32:
-#line 197 "config-parser.rl"
+#line 197 "../../ib/engine/config-parser.rl"
 	{ fsm.te = ( fsm.p)+1;}
 	break;
 	case 33:
-#line 192 "config-parser.rl"
+#line 192 "../../ib/engine/config-parser.rl"
 	{ fsm.te = ( fsm.p);( fsm.p)--;}
 	break;
 	case 34:
-#line 193 "config-parser.rl"
+#line 193 "../../ib/engine/config-parser.rl"
 	{ fsm.te = ( fsm.p);( fsm.p)--;{ { fsm.stack[ fsm.top++] =  fsm.cs;  fsm.cs = 22; goto _again;} }}
 	break;
 	case 35:
-#line 194 "config-parser.rl"
+#line 194 "../../ib/engine/config-parser.rl"
 	{ fsm.te = ( fsm.p);( fsm.p)--;{ { fsm.stack[ fsm.top++] =  fsm.cs;  fsm.cs = 29; goto _again;} }}
 	break;
 	case 36:
-#line 196 "config-parser.rl"
+#line 196 "../../ib/engine/config-parser.rl"
 	{ fsm.te = ( fsm.p);( fsm.p)--;}
 	break;
 	case 37:
@@ -650,7 +650,7 @@ _eof_trans:
 	}
 	}
 	break;
-#line 654 "config-parser.c"
+#line 654 "../../ib/engine/config-parser.c"
 		}
 	}
 
@@ -667,7 +667,7 @@ _again:
 #line 1 "NONE"
 	{ fsm.act = 0;}
 	break;
-#line 671 "config-parser.c"
+#line 671 "../../ib/engine/config-parser.c"
 		}
 	}
 
@@ -687,21 +687,21 @@ _again:
 	while ( __nacts-- > 0 ) {
 		switch ( *__acts++ ) {
 	case 1:
-#line 100 "config-parser.rl"
+#line 100 "../../ib/engine/config-parser.rl"
 	{
         rc = IB_EOTHER;
         ib_log_debug(ib_engine, 4, "ERROR: parser error before \"%.*s\"", (int)(( fsm.p) - mark), mark);
     }
 	break;
 	case 2:
-#line 106 "config-parser.rl"
+#line 106 "../../ib/engine/config-parser.rl"
 	{
         pval = alloc_cpy_marked_string(mark, ( fsm.p), mpcfg);
         ib_list_push(plist, pval);
     }
 	break;
 	case 5:
-#line 122 "config-parser.rl"
+#line 122 "../../ib/engine/config-parser.rl"
 	{
         rc = ib_config_directive_process(cp, dirname, plist);
         if (rc != IB_OK) {
@@ -712,7 +712,7 @@ _again:
         }
     }
 	break;
-#line 716 "config-parser.c"
+#line 716 "../../ib/engine/config-parser.c"
 		}
 	}
 	}
@@ -720,7 +720,7 @@ _again:
 	_out: {}
 	}
 
-#line 259 "config-parser.rl"
+#line 259 "../../ib/engine/config-parser.rl"
 
     /* Ensure that our block is always empty on last chunk. */
     if ( is_last_chunk && blkname != NULL ) {
